More granular spotify test files
Moved Spotify client and query builder to
its own test files. Also created a file with
the raw responses to be imported.
